Use custom renderer when creating embeddings
Rendering makes plugin output available and and handles includes. It
might also help with #15.
The renderer uses markdown like output since all LLMs seem to be very
familiar with it's syntax. This might help them to understand the
document structure better.
This also adds a breadcrumb trail at the top of each chunk which might
help with contextulization as well.