sqlite: avoid warnings on too short vectors
This should not happen in the real world. But when embeddings were
created with a shorter vector model than the model that is used to
embed the query, the cosineSimilarity method threw a whole bunch of
warnings. We now stop the comparison at the vector length.

In the real world the same model for embeddings and the query should be
used, results are unpredictable otherwise. So this is mostly a cosmetic
change for messed up states during development.