Some enhancements on the subsentence splitting
When a sentence is longer than a chunk, it should be split forcefully in
smaller parts - these parts should NOT be the size of a full chunk since
we still want to do some overlap with previous and following texts. I
chose to split into a quarter of a chunk.

This also ensures that whitespace is kept for the split sentences,
because they may be joined with follow up texts.